requirements.txt
openai>=1.0.0
pytest>=7.0.0
pytest-mock>=3.10.0
pytest-asyncio>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.12.0 